refactor(notification): move enuns to common directory
