+++ backend/src/main/java/com/brokendev/backend/common/enums/BoletoPaymentStatus.java
@@ -1,4 +1,4 @@
-package com.brokendev.backend.enums;
+package com.brokendev.backend.common.enums;
 
 public enum BoletoPaymentStatus {
     PENDING,
