+++ backend/src/main/java/com/brokendev/backend/common/enums/PixTransactionStatus.java
@@ -1,4 +1,4 @@
-package com.brokendev.backend.enums;
+package com.brokendev.backend.common.enums;
 
 public enum PixTransactionStatus {
     PENDING,
