+++ backend/src/main/java/com/brokendev/backend/investment/dto/InvestmentRequestDTO.java
@@ -1,6 +1,6 @@
 package com.brokendev.backend.investment.dto;
 
-import com.brokendev.backend.enums.InvestmentType;
+import com.brokendev.backend.common.enums.InvestmentType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotNull;
