+++ backend/src/main/java/com/brokendev/backend/investment/service/InvestmentService.java
@@ -5,10 +5,9 @@ import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.investment.domain.Investment;
 import com.brokendev.backend.investment.dto.InvestmentRequestDTO;
 import com.brokendev.backend.investment.dto.InvestmentResponseDTO;
-import com.brokendev.backend.enums.InvestmentType;
+import com.brokendev.backend.common.enums.InvestmentType;
 import com.brokendev.backend.account.domain.AccountRepository;
 import com.brokendev.backend.investment.domain.InvestmentRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
