+++ backend/src/main/java/com/brokendev/backend/pix_transfer/dto/PixTransferRequestDTO.java
@@ -1,6 +1,6 @@
 package com.brokendev.backend.pix_transfer.dto;
 
-import com.brokendev.backend.enums.PixKeyType;
+import com.brokendev.backend.common.enums.PixKeyType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotNull;
