+++ backend/src/test/java/com/brokendev/backend/account/service/AccountServiceTest.java
@@ -9,10 +9,10 @@ import com.brokendev.backend.boleto_payment.domain.BoletoPayment;
 import com.brokendev.backend.boleto_payment.domain.BoletoPaymentRepository;
 import com.brokendev.backend.common.domain.user.User;
 import com.brokendev.backend.common.domain.user.UserRepository;
+import com.brokendev.backend.common.enums.*;
 import com.brokendev.backend.common.exceptions.InvalidDepositAmountException;
 import com.brokendev.backend.common.exceptions.UserAccountNotFoundException;
 import com.brokendev.backend.investment.domain.Investment;
-import com.brokendev.backend.enums.*;
 import com.brokendev.backend.pix_transfer.domain.PixTransaction;
 import com.brokendev.backend.pix_transfer.domain.PixTransactionRepository;
 import com.brokendev.backend.investment.domain.InvestmentRepository;
