+++ backend/src/test/java/com/brokendev/backend/pix_transfer/service/PixTransferServiceTest.java
@@ -6,8 +6,8 @@ import com.brokendev.backend.common.domain.user.User;
 import com.brokendev.backend.common.exceptions.InsufficientBalanceException;
 import com.brokendev.backend.common.exceptions.PixTransferNotAllowedException;
 import com.brokendev.backend.common.exceptions.UserAccountNotFoundException;
-import com.brokendev.backend.enums.PixKeyType;
-import com.brokendev.backend.enums.PixTransactionStatus;
+import com.brokendev.backend.common.enums.PixKeyType;
+import com.brokendev.backend.common.enums.PixTransactionStatus;
 import com.brokendev.backend.pix_transfer.domain.PixTransaction;
 import com.brokendev.backend.pix_transfer.domain.PixTransactionRepository;
 import com.brokendev.backend.pix_transfer.dto.PixTransferRequestDTO;
